+++ tests/unit/test_main_process.py
@@ -3,7 +3,7 @@ import pytest
 import sys
 import yaml
 
-from loadEnv import loadEnvFile
+from load_env import load_env_file
 from main import main, registerProcesses, createArgParser
 import processes
 
@@ -114,13 +114,13 @@ class TestMainProcess:
 
     def test_loadEnvFile_default_file(self, sampleEnvFile, mocker):
         mockStream = mocker.MagicMock()
-        mockOpen = mocker.patch('loadEnv.open')
+        mockOpen = mocker.patch('load_env.open')
         mockOpen.return_value = mockStream
         mockYaml = mocker.patch('yaml.full_load')
         mockYaml.return_value = sampleEnvFile
         mockEnviron = mocker.patch.dict('os.environ', {})
 
-        loadEnvFile('test', None)
+        load_env_file('test', None)
 
         mockOpen.assert_called_with('local.yaml')
 
@@ -129,13 +129,13 @@ class TestMainProcess:
         
     def test_loadEnvFile_specified_file(self, sampleEnvFile, mocker):
         mockStream = mocker.MagicMock()
-        mockOpen = mocker.patch('loadEnv.open')
+        mockOpen = mocker.patch('load_env.open')
         mockOpen.return_value = mockStream
         mockYaml = mocker.patch('yaml.full_load')
         mockYaml.return_value = sampleEnvFile
         mockEnviron = mocker.patch.dict('os.environ', {})
 
-        loadEnvFile('test', './test/path/{}.yaml')
+        load_env_file('test', './test/path/{}.yaml')
 
         mockOpen.assert_called_with('./test/path/test.yaml')
 
@@ -143,13 +143,13 @@ class TestMainProcess:
         assert mockEnviron['TEST_PORT'] == '9999'
 
     def test_loadEnvFile_missing_file(self, sampleEnvFile, mocker):
-        mockOpen = mocker.patch('loadEnv.open')
+        mockOpen = mocker.patch('load_env.open')
         mockOpen.side_effect = FileNotFoundError
         mockYaml = mocker.patch('yaml.full_load')
         mockEnviron = mocker.patch.dict('os.environ', {})
 
         with pytest.raises(FileNotFoundError):
-            loadEnvFile('test', None)
+            load_env_file('test', None)
             mockYaml.assert_not_called
 
         assert mockEnviron.get('TEST_NAME', None) is None
@@ -157,14 +157,14 @@ class TestMainProcess:
 
     def test_loadEnvFile_invalid_yaml(self, sampleEnvFile, mocker):
         mockStream = mocker.MagicMock()
-        mockOpen = mocker.patch('loadEnv.open')
+        mockOpen = mocker.patch('load_env.open')
         mockOpen.return_value = mockStream
         mockYaml = mocker.patch('yaml.full_load')
         mockYaml.side_effect = yaml.YAMLError
         mockEnviron = mocker.patch.dict('os.environ', {})
 
         with pytest.raises(yaml.YAMLError):
-            loadEnvFile('test', None)
+            load_env_file('test', None)
 
         assert mockEnviron.get('TEST_NAME', None) is None
         assert mockEnviron.get('TEST_PORT', None) is None